added the rest of the messages api
